pruaba 19

+++ src/exercises.cpp
@@ -26,9 +26,9 @@ void exercise_2(double A, double B, double C) {
     A;
     B;
     C;
-    double  X1,X2;
+    float  X1,X2;
 
-    double x = (B*B) - (4*A*C);
+    float x = (B*B) - (4*A*C);
     
     if(x<=0)
     {
